requirements.txt
google-adk>=0.3.0
fastapi>=0.115.0
uvicorn>=0.30.0
mcp[cli]
httpx>=0.27.0
python-dotenv>=1.0.0
